Rework extract test, move the comparison logic out
